pawnwalker/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
Django==2.2.10
django-booking==0.7.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-countries==6.0
django-datetime-widget==0.9.3
django-hvad==1.8.0
django-libs==2.0.2
django-reservation==0.2.9
filelock==3.0.12
googlemaps==4.2.0
idna==2.9
importlib-metadata==1.5.0
locationiq==0.0.2
pipenv==2018.11.26
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==20.0.7
virtualenv-clone==0.5.3
zipp==3.0.0
